# Use platform setjmp/longjmp to optimize function calls