--- a/lib/api/src/native.rs
+++ b/lib/api/src/native.rs
@@ -137,6 +137,7 @@ macro_rules! impl_native_traits {
                         };
                         unsafe {
                             wasmer_vm::wasmer_call_trampoline(
+                                &self.store,
                                 self.vmctx(),
                                 trampoline,
                                 self.address(),