--- a/lib/vm/build.rs
+++ b/lib/vm/build.rs
@@ -3,16 +3,17 @@
 use std::env;
 
 fn main() {
-    println!("cargo:rerun-if-changed=src/trap/helpers.c");
+    println!("cargo:rerun-if-changed=src/trap/handlers.c");
+
     cc::Build::new()
         .warnings(true)
         .define(
             &format!(
-                "CFG_TARGET_OS_{}",
+                "TARGET_OS_{}",
                 env::var("CARGO_CFG_TARGET_OS").unwrap().to_uppercase()
             ),
             None,
         )
-        .file("src/trap/helpers.c")
-        .compile("helpers");
+        .file("src/trap/handlers.c")
+        .compile("handlers");
 }