--- a/lib/vm/src/trap/helpers.c
+++ b/lib/vm/src/trap/helpers.c
@@ -3,12 +3,26 @@
 
 #include <setjmp.h>
 
+// Note that `sigsetjmp` and `siglongjmp` are used here where possible to
+// explicitly pass a 0 argument to `sigsetjmp` that we don't need to preserve
+// the process signal mask. This should make this call a bit faster b/c it
+// doesn't need to touch the kernel signal handling routines.
+#ifdef CFG_TARGET_OS_WINDOWS
+#define platform_setjmp(buf) setjmp(buf)
+#define platform_longjmp(buf, arg) longjmp(buf, arg)
+#define platform_jmp_buf jmp_buf
+#else
+#define platform_setjmp(buf) sigsetjmp(buf, 0)
+#define platform_longjmp(buf, arg) siglongjmp(buf, arg)
+#define platform_jmp_buf sigjmp_buf
+#endif
+
 int RegisterSetjmp(
     void **buf_storage,
     void (*body)(void*),
     void *payload) {
-  jmp_buf buf;
-  if (setjmp(buf) != 0) {
+  platform_jmp_buf buf;
+  if (platform_setjmp(buf) != 0) {
     return 0;
   }
   *buf_storage = &buf;
@@ -17,6 +31,6 @@ int RegisterSetjmp(
 }
 
 void Unwind(void *JmpBuf) {
-  jmp_buf *buf = (jmp_buf*) JmpBuf;
-  longjmp(*buf, 1);
+  platform_jmp_buf *buf = (platform_jmp_buf*) JmpBuf;
+  platform_longjmp(*buf, 1);
 }