--- a/lib/vm/src/trap/mod.rs
+++ b/lib/vm/src/trap/mod.rs
@@ -3,12 +3,13 @@
 
 //! This is the module that facilitates the usage of Traps
 //! in Wasmer Runtime
-mod trapcode;
-mod traphandlers;
 
-pub use trapcode::TrapCode;
-pub use traphandlers::{
+mod code;
+mod handlers;
+
+pub use code::TrapCode;
+pub use handlers::{
     catch_traps, catch_traps_with_result, raise_lib_trap, raise_user_trap, wasmer_call_trampoline,
     Trap,
 };
-pub use traphandlers::{init_traps, resume_panic};
+pub use handlers::{init_traps, resume_panic};