… CHANGELOG.md
… Cargo.lock
… lib/vm/Cargo.toml
… lib/vm/src/trap/code.rs
… lib/vm/src/trap/handlers/macos.rs
… lib/vm/src/trap/handlers/mod.rs
… lib/vm/src/trap/handlers/unix.rs
… lib/vm/src/trap/handlers/windows.rs
… lib/vm/src/vmcontext.rs