@syrusakbary 2021-03-05
@Hywan I'm trying to run this locally and I'm getting this in x86 (getting somethings similar in arm after hacking to get things running):

```
test wast::cranelift::spec::call_indirect ... test wast::cranelift::spec::call_indirect has been running for over 60 seconds
test wast::cranelift::spec::call ... test wast::cranelift::spec::call has been running for over 60 seconds
test wast::cranelift::spec::fac ... test wast::cranelift::spec::fac has been running for over 60 seconds
test wast::cranelift::spec::multi_value::call ... test wast::cranelift::spec::multi_value::call has been running for over 60 seconds
test wast::cranelift::spec::multi_value::call_indirect ... test wast::cranelift::spec::multi_value::call_indirect has been running for over 60 seconds
test wast::cranelift::spec::multi_value::fac ... test wast::cranelift::spec::multi_value::fac has been running for over 60 seconds
test wast::cranelift::wasmer::fac ... test wast::cranelift::wasmer::fac has been running for over 60 seconds
test wast::cranelift::wasmer::stack_overflow ... test wast::cranelift::wasmer::stack_overflow has been running for over 60 seconds
```